+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 * **@dlr-eoc/map-ol:**
   - Styles from OpenLayers are no longer included in this component [e3e98c1](https://github.com/dlr-eoc/ukis-frontend-libraries/pull/151/commits/e3e98c1a9a04a23c2109a509e902d14c1b6bdc9d). You have to add them to your application manually see [`@dlr-eoc/map-ol/README.md`](projects/map-ol/README.md).
+### Features
+* **@dlr-eoc/map-ol:**
+ - The function `zoomInOut` now accepts an optional object with `callback`, `duration`, `easing` and `zoomStep`.
 
 ### Bug Fixes
 * **@dlr-eoc/map-ol:**